fix(swagger): Exclude MetaClass from swagger model generation (#501)
* fix(swagger): Exclude MetaClass from swagger model generation

All groovy classes have an implicit metaClass field that handles
metaprogramming. In cases where a groovy class is used in an
API the springfox library to generate Swagger docs is including
the metaClass in the model for the class. In addition to being
confusing (as a producer/consumer of the API would never set
that field), it also leads to terrible performance in some
cases.

The performance issues come from the fact that it then tries to
build a model of MetaClass by reflecting its fields, etc. until
it has built models of most of the Groovy classes.

In particular, before this change, front50 took 140 seconds to start;
with this change it now takes 13 seconds.

* fix(swagger): Make groovy implementation dependency

This does not need to be an API dependency.

* fix(swagger): Only bring in core groovy

We don't need all of it, only the core

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>