fix(docker): Fix docker registry health check (#4354)
* test(kubernetes): Add tests to KubernetesHealthIndicator

Before refactoring this class to push some logic to the base class,
add some tests on it.  This also adds a couple of useful stub classes
to make it easier to create stub accounts.

* refactor(kubernetes): Push logic for kubernetes health to base

The logic for checking the health of a Kubernetes account is similar
to what we want for docker accounts (and likely for other accounts).
Namely, we want to report errors with an individual account, but don't
want to bring down all of clouddriver because of failures with
individual accounts. Make a base class that can be extended by cloud
providers to implement this functionality.

This also makes one small change to the way unhealthy accounts are
reported in detailed health check results; as the accounts will already
appear under a key of 'kubernetes', I've eliminated the redundant prefix of
'kubernetes:' from each of the account names.

* feat(kubernetes): Add metrics to kubernetes health check

Add support to the base AccountHealthIndicator to report the
number of unhealthy accounts, which then adds support for this
to the Kubernetes health indicator.

* fix(docker): Fix docker registry health check

The docker registry health check currently fails if any single docker
account is unreachable.  (Though it will not fail after startup if it
has passed once.)

This means that a single bad account can cause all of clouddriver to
fail to start, which is bad for failure isolation. Rewrite the health
check to use the new AccountHealthIndicator class which will report
errors (both as metrics and in the detailed health view) but will not
cause clouddriver to report itself as DOWN.

* fix(kubernetes): Fix tests depending on map iteration order

When applied to a Map, containsExactly requires that iterating over
the map returns the supplied entries in order. As we are not using
an ordered map, this will lead to intermittent test failures.

Replace this with containsOnly, which asserts the map contains
exactly the supplied entries, in any order.

Also fix a copyright header.

* style(docker): Address PR review comments

* style(docker): Fix remaining PR comment