refactor(artifacts): Make artifact classes final and package-private (#4366)
* refactor(artifacts): Compile clouddriver-artifacts with Java compiler

This package is already entirely in java, so no additional changes
are needed other than setting the flag to compile it with the java
compiler.

We can also remove the useJUnitPlatform() setting as it has now been
added to the base clouddriver build file so does not need to be
specified in individual modules.

* refactor(artifacts): Make artifact classes final and package-private

In general, the interface of the artifacts module is to allow clients
to pass in an artifact and get back an InputStream of the resolved
artifact's contents. All of the accounts, configuration, and credentials
are just implementation details of how this module does that and should
be made package-private (and final as they were not designed for
inheritance).

The interfaces of course remain public, but clients should not need
to know which particular implementation of an artifact account is
handling their request.  In general, this is the case, though there
is one place that breaks this encapsulation, where the cloud foundry
deploy depends on details of the Maven artifact resolver. For that
reason (to keep this refactor smaller) I'm just leaving that resolver
public but ideally we'd find a better way of supporting what that
consumer needs rather than make the entire resolver public.

* refactor(artifacts): Remove GCE reference from ArtifactUtils

There is a GCE-specific string contstant in ArtifactUtils; move it
to a more appropriate location in the GCE module. (This in particular
should not live in the core artifacts module as we don't even have
a gce/image artifact handler at all.)

* refactor(artifacts): Move ArtifactUtils to appengine

The ArtifactUtils class is only used for appengine-specific operations
of un-taring an archive. Rather than have it live in the core module,
move it to the appengine module near its only use case.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>