refactor(core): Push halyard defaults into base config defaults (#4368)
* fix(redis): Clean up cache hashing config

The config setting caching.redis.hashingEnabled controls whether
we try to short-circuit sending data to redis by comparing a hash
of the data client-side with a hash of the data server-side.

This setting is on by default, (see DEFAULT_HASHING_ENABLED in
RedisCacheOptions.java) so there's no need to explicitly set it
in config files.

Also remove the setting turning it off in the HA clouddriver instances;
this has no effect for -ro and -ro-deck as those instances never
write to the cache anyway. This will be a change for the -rw instance,
but in general this should improve performance. (At least there's no
reason a priori to have the default different between different
clouddriver instances.)

* refactor(core): Remove explicit setting of shutdownWaitSeconds

This parameter controls how long clouddriver will wait for in-flight
tasks before shutting down. It is defaulted to 600 for all users
deploying with halyard (which is most users). Rather than put this
setting directly in the halyard-generated config, set the default
in code.

This will change the default for users that don't use Halyard
(primarily Netflix). I'm happy to pick a different default if that
is more reasonable, but given this is what all Halyard users have
had for years it seems like the smallest change.

* refactor(core): Push defaulting of port and host to base config

Halyard has been using the settings in spinnaker.yml to default
the port and host for clouddriver for all Halyard users by adding
a block to the user's config.

Push this logic to the base clouddriver config; the defaults are
the same for anyone that does not have a services.* block in their
config overriding the values, but for users who do (all Halyard users)
this now respects those overrides.

* refactor(core): Add README to halconfig directory

Let's note that the configs in this directory are deprecated and
should not be updated. (I didn't add the comment to the files
themselves as they get interpolated, including comments, into the
users' config files.)

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>