chore(mergify): Allow Travis success on release branches (#4463)
Mergify is not working on the release branches because it requires
the Github actions build to return success, but we only switched
to using Github actions after cutting the 1.19 branch.

To remedy this, add another rule that allows merging if Travis CI
passes on a branch.  (There's no way to make an OR condition within
a rule, so we need to add another rule.) This way, we'll merge if
either Github actions or Travis CI reports success.  Once the
branches running Travis are no longer maintained (when Spinnaker
1.23 is released) we can delete this rule.