refactor(kubernetes): Clean up metrics code (#4472)
* test(kubernetes): Add some tests to metric fetching logic

Before making a few changes to the call to get pod metrics, add some
tests around this to ensure the changes don't break existing
functionality.

* refactor(kubernetes): Clean up metrics code

This started out just trying to remove the noisy log statement
"No output from `kubectl top` command, no metrics to report",
which is logged with severity warn any time we cache an empty
namespace. There's no need to log this as it's a perfectly normal
situation; we asked for pod metrics and there were none.

But then I noticed some things that could be cleaned up a bit 😄.
So this PR does the following:
* Extracts the logic to parse the output of kubectl top out of
the already large job executor class and into a helper class
* Replaces use of .split() which compiles a regex every call and
requires us to deal in arrays with Guava Splitters
* Cleans up KubernetesPodMetric, making things immutable and also
removing an excessive buidler from ContainerMetric which has only
two fields of different types.

* refactor(kubernetes): Better encapsulate metric parsing

If we have parseMetrics just return a ContainerMetric directly, we
can keep its helper class MetricLine private, and avoid having the
consumer need to know about it.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>