feat(kubernetes): Apply manifests with generateName using create (#4492)
* feat(kubernetes): Add support for kubectl create

Some manifest need to be applied with kubectl create; one example
is manifests that specify a generateName instead of a name. In this
case, we can't apply or replace, as both of these operations require
looking up the existing manifest which is impossible when the name
changes each time.

This commit adds support to KubernetesV2Credentials to create a
manifest using kubectl create operation.

* test(kubernetes): Add tests to CanDeploy

Before lightly refactoring and adding some new functionality to
the CanDeploy interface, add some tests to validate the existing
behavior.

* refactor(kubernetes): Use switch in CanDeploy

The logic in CanDeploy is a bit too complicated; let's make it
cleaner by using a switch over the enum of possible deploy
strategies.

* feat(kubernetes): Apply manifests with generateName using create

Manifests that have a generateName instead of a name need to be
created using a create operation, as other operations require
looking up the existing manifest to reconcile changes, but this
is not possible if we don't have the manifest's name.

This change updates CanDeploy to check if the manifest has a
generateName (and no name, as this would take priority), and to
always use 'kubectl create' if that is the case.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>